+++ app/Services/Api/Admin/ConcertService.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Services\Api\Main;
+namespace App\Services\Api\Admin;
 
 use App\Services\ApiService;
-use App\Http\Resources\Main\ConcertResource;
+use App\Http\Resources\Admin\ConcertResource;
 
 class ConcertService extends ApiService
 {
@@ -26,6 +26,18 @@ class ConcertService extends ApiService
     }
 
     /**
+     * Store function.
+     * 
+     * @param $request
+     */
+    public function store($request)
+    {
+        $this->concertInterface->create($request);
+
+        return $this->index();
+    }
+
+    /**
      * Show function.
      * 
      * @param $path
@@ -37,5 +49,36 @@ class ConcertService extends ApiService
         return $this->createResponse(trans('api.response.accepted'), [
             'data' => new ConcertResource($concert)
         ], 206);
+    }
+
+    /**
+     * Update function.
+     * 
+     * @param $request
+     * @param $id
+     */
+    public function update($request, $id)
+    {
+        $this->concertInterface->update(intval($id), $request);
+
+        if (empty($request)) {
+            return $this->createResponse(trans('api.response.updated'), [
+                'data' => trans('api.response.no_data_changed')
+            ], 202);
+        }
+
+        return $this->index();
+    }
+
+    /**
+     * Destroy function.
+     * 
+     * @param $id
+     */
+    public function destroy($id)
+    {
+        $this->concertInterface->deleteById($id);
+
+        return $this->index();
     }
 }
